requirements_.txt
aniso8601==9.0.1
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
coverage==5.3
Faker==13.2.0
Flask==2.3.2
Flask-Cors==4.0.0
Flask-RESTful==0.3.10
flask-sqlalchemy==3.0.5
greenlet==2.0.2
gunicorn==21.2.0
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.1
psycopg2==2.9.6
pyparsing==3.0.7
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.2
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
six==1.16.0
SQLAlchemy==1.4.31
urllib3==2.0.4
Werkzeug==2.3.6
zipp==3.16.2
